10592bytes

+++ build.sh
@@ -1,14 +1,27 @@
-curl "http://crystal-lang.s3.amazonaws.com/pcl/libpcl-1.12-1-linux-x86_64.tar.gz" | tar xz -C /tmp
-git clone git://github.com/ivmai/libatomic_ops.git
-git clone git://github.com/ivmai/bdwgc.git
-ln -s `pwd`/libatomic_ops `pwd`/bdwgc/libatomic_ops
-cd bdwgc
-autoreconf -vif
-automake --add-missing
-./configure
-make
-cd ..
-export LIBRARY_PATH=/tmp/libpcl-1.12-1/lib:./bdwgc/.libs
-crystal build hello.cr --release --emit obj
-gcc hello.o -o hello -s -l:libpcre.a -l:libgc.a -l:libevent.a -l:libpcl.a -lpthread -lunwind
+#!/bin/bash
+
+dirname=ivmai
+
+set -e
+
+crystal -v
+echo
+
+set -x
+
+if [ -e $dirname ]; then
+  git clone git://github.com/ivmai/libatomic_ops.git
+  git clone git://github.com/ivmai/bdwgc.git
+  ln -s `pwd`/libatomic_ops `pwd`/bdwgc/libatomic_ops
+  cd bdwgc
+  autoreconf -vif
+  automake --add-missing
+  ./configure
+  make
+  cd ..
+fi
+
+export LIBRARY_PATH=./bdwgc/.libs
+crystal build hello.cr --emit obj --prelude=empty --release
+gcc hello.o -o hello -s -lgc -lpthread -lunwind
 wc -c < hello
